fix(review): address CodeRabbit feedback on PR #1314
- orchestrator/prompt.md: replace `integrations_agent` references in
  the decision tree with the actual synthesised `delegate_{toolkit}` /
  `delegate_run_code` / `delegate_plan` etc. tool names so the prompt
  matches what the LLM sees in its function-calling schema
- tests/agent_retrieval_e2e.rs: strengthen memory_tree tool assertion to
  use exact line matching (`"memory_tree"` / `"memory_tree",`) rather
  than a substring match that could hit comments or prefixed names
- subagent_runner/ops_truncation_tests.rs: extract the three
  max_result_chars truncation tests into their own file, keeping
  ops_tests.rs under the 500-line guideline threshold
