+++ src/openhuman/agent/agents/orchestrator/prompt.md
@@ -21,11 +21,12 @@ Follow this sequence for every user message:
    - Yes: use direct tools first (`current_time`, `cron_*`, `memory_*`, `composio_list_connections`, etc.).
    - No: continue.
 3. **Does this need specialised execution?**
-   - If external SaaS integration work is required, delegate to `integrations_agent` with the right toolkit.
-   - If code writing/execution/debugging is required, delegate to `code_executor`.
-   - If web/doc crawling is required, delegate to `researcher`.
-   - If complex multi-step decomposition is required, delegate to `planner` (and only then route deeper if necessary).
-   - If code review is requested, delegate to `critic`.
+   - If external SaaS integration work is required, use `delegate_{toolkit}` (e.g. `delegate_gmail`, `delegate_notion`).
+   - If code writing/execution/debugging is required, use `delegate_run_code`.
+   - If web/doc crawling is required, use `delegate_researcher`.
+   - If complex multi-step decomposition is required, use `delegate_plan`.
+   - If code review is requested, use `delegate_critic`.
+   - If memory archiving or distillation is required, use `delegate_archivist`.
 4. **After delegation**, summarise results clearly and concisely.
 
 Default bias: **do not spawn a sub-agent when a direct response or direct tool call is sufficient**.
